# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AzureCertify AI Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exams/configure:
    post:
      operationId: configureExam
      tags: [exams]
      summary: Configure and start an exam session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamConfig"
      responses:
        "200":
          description: Exam session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamSession"
  /exams/sessions/{sessionId}:
    get:
      operationId: getExamSession
      tags: [exams]
      summary: Get exam session with questions
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exam session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamSession"
  /exams/sessions/{sessionId}/submit:
    post:
      operationId: submitExam
      tags: [exams]
      summary: Submit exam answers
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamSubmission"
      responses:
        "200":
          description: Exam results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
...
components:
  schemas:
...
    ExamConfig:
      type: object
      properties:
        certificationCode:
          type: string
          enum: [AZ-900, AZ-104, AZ-305]
        mode:
          type: string
          enum: [practice, exam]
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        questionCount:
          type: integer
          minimum: 10
          maximum: 60
        userId:
          type: string
      required: [certificationCode, mode, difficulty, questionCount]
    ExamQuestion:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              text:
                type: string
            required: [key, text]
        correctAnswer:
          type: string
        explanation:
          type: string
        topic:
          type: string
        difficulty:
          type: string
        isAiGenerated:
          type: boolean
        validationScore:
          type: number
      required: [id, content, options, correctAnswer, explanation, topic, difficulty, isAiGenerated]
    ExamSession:
      type: object
      properties:
        sessionId:
          type: string
        certificationCode:
          type: string
        mode:
          type: string
        difficulty:
          type: string
        questionCount:
          type: integer
        timeLimitMinutes:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ExamQuestion"
        createdAt:
          type: string
      required: [sessionId, certificationCode, mode, difficulty, questionCount, timeLimitMinutes, questions, createdAt]
    ExamSubmission:
      type: object
      properties:
        answers:
          type: object
          additionalProperties:
            type: string
        timeSpentSeconds:
          type: integer
      required: [answers, timeSpentSeconds]
    ExamResult:
      type: object
      properties:
        attemptId:
          type: string
        score:
          type: number
        passed:
          type: boolean
        correctCount:
          type: integer
        totalCount:
          type: integer
        timeSpentSeconds:
          type: integer
        questionResults:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              correct:
                type: boolean
              userAnswer:
                type: string
              correctAnswer:
                type: string
              explanation:
                type: string
              topic:
                type: string
            required: [questionId, correct, userAnswer, correctAnswer, explanation, topic]
        topicBreakdown:
          type: array
          items:
            type: object
            properties:
              topic:
                type: string
              correct:
                type: integer
              total:
                type: integer
              percentage:
                type: number
            required: [topic, correct, total, percentage]
      required: [attemptId, score, passed, correctCount, totalCount, timeSpentSeconds, questionResults, topicBreakdown]
...