# lib/api-zod/src/generated/types/examConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AzureCertify AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { ExamConfigCertificationCode } from "./examConfigCertificationCode";
import type { ExamConfigDifficulty } from "./examConfigDifficulty";
import type { ExamConfigMode } from "./examConfigMode";

export interface ExamConfig {
  certificationCode: ExamConfigCertificationCode;
  mode: ExamConfigMode;
  difficulty: ExamConfigDifficulty;
  /**
   * @minimum 10
   * @maximum 60
   */
  questionCount: number;
  userId?: string;
}