# lib/api-zod/src/generated/types/examSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AzureCertify AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { ExamQuestion } from "./examQuestion";

export interface ExamSession {
  sessionId: string;
  certificationCode: string;
  mode: string;
  difficulty: string;
  questionCount: number;
  timeLimitMinutes: number;
  questions: ExamQuestion[];
  createdAt: string;
}