# lib/api-zod/src/generated/types/examSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AzureCertify AI Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { ExamSubmissionAnswers } from "./examSubmissionAnswers";

export interface ExamSubmission {
  answers: ExamSubmissionAnswers;
  timeSpentSeconds: number;
}